refactor: split oversized gateway_error files by extracting tests (#221)
Extract inline test modules from conversions.rs (1295 lines) and
response.rs (886 lines) into separate files using #[path] attribute.

- conversions.rs: split tests into 3 files by error source
  (provider, a2a, mcp) — each under 410 lines
- response.rs: split tests into 2 files by focus area
  (structs/integration, status codes) — each under 460 lines
- Both source files now under 360 lines

No public API changes. All 224 gateway_error tests pass.
