+++ src/utils/error/gateway_error/response_tests_status.rs
@@ -1,0 +1,459 @@
+use super::*;
+use crate::core::providers::unified_provider::ProviderError;
+use actix_web::http::StatusCode;
+
+// ==================== GatewayError ResponseError Tests ====================
+
+#[test]
+fn test_gateway_error_config_response() {
+    let error = GatewayError::Config("Invalid configuration".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[test]
+fn test_gateway_error_internal_response() {
+    let error = GatewayError::Internal("Internal server error".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[test]
+fn test_gateway_error_auth_response() {
+    let error = GatewayError::Auth("Invalid token".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+}
+
+#[test]
+fn test_gateway_error_forbidden_response2() {
+    let error = GatewayError::Forbidden("Permission denied".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::FORBIDDEN);
+}
+
+#[test]
+fn test_gateway_error_rate_limit_response() {
+    let error = GatewayError::RateLimit {
+        message: "Rate limit exceeded".to_string(),
+        retry_after: Some(60),
+        rpm_limit: Some(100),
+        tpm_limit: Some(50000),
+    };
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::TOO_MANY_REQUESTS);
+    assert_eq!(
+        response
+            .headers()
+            .get("Retry-After")
+            .unwrap()
+            .to_str()
+            .unwrap(),
+        "60"
+    );
+    assert_eq!(
+        response
+            .headers()
+            .get("X-RateLimit-Limit-Requests")
+            .unwrap()
+            .to_str()
+            .unwrap(),
+        "100"
+    );
+    assert_eq!(
+        response
+            .headers()
+            .get("X-RateLimit-Limit-Tokens")
+            .unwrap()
+            .to_str()
+            .unwrap(),
+        "50000"
+    );
+}
+
+#[test]
+fn test_gateway_error_validation_response() {
+    let error = GatewayError::Validation("Invalid input".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_gateway_error_not_found_response() {
+    let error = GatewayError::NotFound("Resource not found".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_FOUND);
+}
+
+#[test]
+fn test_gateway_error_conflict_response() {
+    let error = GatewayError::Conflict("Resource conflict".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::CONFLICT);
+}
+
+#[test]
+fn test_gateway_error_bad_request_response() {
+    let error = GatewayError::BadRequest("Bad request".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_gateway_error_timeout_response() {
+    let error = GatewayError::Timeout("Request timeout".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::REQUEST_TIMEOUT);
+}
+
+#[test]
+fn test_gateway_error_provider_unavailable_response() {
+    let error = GatewayError::Unavailable("Provider down".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_gateway_error_circuit_breaker_response() {
+    let error = GatewayError::Unavailable("Circuit open".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_gateway_error_network_response() {
+    let error = GatewayError::Network("Network error".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+#[test]
+fn test_gateway_error_parsing_response() {
+    let error = GatewayError::Validation("Parse error".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_gateway_error_not_implemented_response() {
+    let error = GatewayError::NotImplemented("Feature not implemented".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_IMPLEMENTED);
+}
+
+#[test]
+fn test_gateway_error_forbidden_response() {
+    let error = GatewayError::Forbidden("Forbidden".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::FORBIDDEN);
+}
+
+#[test]
+fn test_gateway_error_external_response() {
+    let error = GatewayError::Network("External service error".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+#[test]
+fn test_gateway_error_no_providers_available_response() {
+    let error = GatewayError::Unavailable("No providers".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_gateway_error_provider_not_found_response() {
+    let error = GatewayError::NotFound("Provider not found".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_FOUND);
+}
+
+#[test]
+fn test_gateway_error_no_providers_for_model_response() {
+    let error = GatewayError::BadRequest("No providers for model".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_gateway_error_no_healthy_providers_response() {
+    let error = GatewayError::Unavailable("No healthy providers".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+// ==================== Provider Error Tests ====================
+
+#[test]
+fn test_provider_error_rate_limit_response() {
+    let provider_error = ProviderError::RateLimit {
+        provider: "openai",
+        message: "Rate limit exceeded".to_string(),
+        retry_after: Some(60),
+        rpm_limit: Some(100),
+        tpm_limit: Some(10000),
+        current_usage: Some(0.95),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::TOO_MANY_REQUESTS);
+}
+
+#[test]
+fn test_provider_error_quota_exceeded_response() {
+    let provider_error = ProviderError::QuotaExceeded {
+        provider: "anthropic",
+        message: "Monthly quota exceeded".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::PAYMENT_REQUIRED);
+}
+
+#[test]
+fn test_provider_error_model_not_found_response() {
+    let provider_error = ProviderError::ModelNotFound {
+        provider: "openai",
+        model: "gpt-5".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_FOUND);
+}
+
+#[test]
+fn test_provider_error_invalid_request_response() {
+    let provider_error = ProviderError::InvalidRequest {
+        provider: "openai",
+        message: "Invalid parameters".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_provider_error_timeout_response() {
+    let provider_error = ProviderError::Timeout {
+        provider: "openai",
+        message: "Request timed out after 30s".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::GATEWAY_TIMEOUT);
+}
+
+#[test]
+fn test_provider_error_unavailable_response() {
+    let provider_error = ProviderError::ProviderUnavailable {
+        provider: "azure",
+        message: "Service maintenance".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_provider_error_authentication_response() {
+    let provider_error = ProviderError::Authentication {
+        provider: "openai",
+        message: "Invalid API key".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+}
+
+#[test]
+fn test_provider_error_api_error_passthrough() {
+    let provider_error = ProviderError::ApiError {
+        provider: "unknown",
+        status: 418,
+        message: "I'm a teapot".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status().as_u16(), 418);
+}
+
+#[test]
+fn test_provider_error_api_error_invalid_status_fallback() {
+    let provider_error = ProviderError::ApiError {
+        provider: "unknown",
+        status: 0,
+        message: "Invalid status".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+#[test]
+fn test_provider_error_context_length_response() {
+    let provider_error = ProviderError::ContextLengthExceeded {
+        provider: "openai",
+        max: 8192,
+        actual: 10000,
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_provider_error_not_supported_response() {
+    let provider_error = ProviderError::NotSupported {
+        provider: "openai",
+        feature: "vision".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_IMPLEMENTED);
+}
+
+#[test]
+fn test_provider_error_configuration_response() {
+    let provider_error = ProviderError::Configuration {
+        provider: "azure",
+        message: "Missing deployment".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[test]
+fn test_provider_error_deployment_response() {
+    let provider_error = ProviderError::DeploymentError {
+        provider: "azure",
+        deployment: "gpt-4-east".to_string(),
+        message: "Deployment not found".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::NOT_FOUND);
+}
+
+#[test]
+fn test_provider_error_routing_response() {
+    let provider_error = ProviderError::RoutingError {
+        provider: "openrouter",
+        attempted_providers: vec!["openai".to_string(), "anthropic".to_string()],
+        message: "All providers failed".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_provider_error_other_variant_response() {
+    let provider_error = ProviderError::Other {
+        provider: "custom",
+        message: "Unknown error".to_string(),
+    };
+    let error = GatewayError::Provider(provider_error);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+// ==================== Previously-wildcard Variant Tests ====================
+
+#[test]
+fn test_gateway_error_jwt_response() {
+    use jsonwebtoken::{errors::Error as JwtError, errors::ErrorKind};
+    let jwt_error = JwtError::from(ErrorKind::InvalidToken);
+    let error = GatewayError::Auth(format!("JWT error: {}", jwt_error));
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+}
+
+#[test]
+fn test_gateway_error_serialization_response() {
+    let json_err: serde_json::Error = serde_json::from_str::<serde_json::Value>("bad").unwrap_err();
+    let error: GatewayError = json_err.into();
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[test]
+fn test_gateway_error_cache_response() {
+    let error = GatewayError::Storage("Cache unavailable".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_gateway_error_crypto_response() {
+    let error = GatewayError::Auth("Encryption failed".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+}
+
+#[test]
+fn test_gateway_error_file_storage_response() {
+    let error = GatewayError::Internal("Write failed".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[test]
+fn test_gateway_error_http_client_response() {
+    let req_err = reqwest::Client::new()
+        .get("not-a-valid-url")
+        .build()
+        .unwrap_err();
+    let error = GatewayError::HttpClient(req_err);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+#[test]
+fn test_gateway_error_io_response() {
+    let io_err = std::io::Error::new(std::io::ErrorKind::PermissionDenied, "permission denied");
+    let error = GatewayError::Io(io_err);
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}
+
+#[test]
+fn test_gateway_error_vector_db_response() {
+    let error = GatewayError::Storage("Vector search failed".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[test]
+fn test_gateway_error_yaml_response() {
+    let yaml_err: serde_yml::Error =
+        serde_yml::from_str::<serde_yml::Value>("key: [unclosed").unwrap_err();
+    let error = GatewayError::Serialization(yaml_err.to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+}
+
+#[cfg(feature = "vector-db")]
+#[test]
+fn test_gateway_error_qdrant_response() {
+    let error = GatewayError::Storage("Qdrant unavailable".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
+
+#[cfg(feature = "websockets")]
+#[test]
+fn test_gateway_error_websocket_response() {
+    let error = GatewayError::Network("WS connection failed".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
+}
+
+#[cfg(feature = "s3")]
+#[test]
+fn test_gateway_error_s3_response() {
+    let error = GatewayError::Storage("Bucket not found".to_string());
+    let response = error.error_response();
+    assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
+}
